+++ frontend-client/src/views/AdminPanel.vue
@@ -22,8 +22,8 @@
         <HeaderPanel
           name="Admin - Pending Reservations"
           :filters="filters"
-          :fetchData="fetchData"
-          :clearFilter="clearFilter"
+          :fetchData="requestData"
+          :clearFilters="clearFilter"
           @inputUpdate="filters['global'].value = $event" />
       </template>
       <template #empty v-if="!loading">No items found.</template>
@@ -39,7 +39,7 @@
         <template #filter="{ filterModel }">
           <span class="p-input-icon-left">
             <i class="pi pi-search" />
-            <AutoComplete v-model="filterModel.value" :suggestions="filteredItems" @complete="searchItems" :virtualScrollerOptions="{ itemSize: 38, style: 'overflow-x: hidden' }" dropdown />
+            <AutoComplete v-model="filterModel.value" :suggestions="filteredItems" @complete="search" :virtualScrollerOptions="{ itemSize: 38, style: 'overflow-x: hidden' }" dropdown />
           </span>
         </template>
       </Column>
@@ -147,13 +147,19 @@ import MultiSelect from 'primevue/multiselect';
 import Calendar from 'primevue/calendar';
 import Toast from 'primevue/toast';
 import ImageColumn from '@/components/inventory-list/ImageColumn.vue';
-import { FilterMatchMode, FilterOperator } from 'primevue/api';
 import { useRouter } from 'vue-router';
 import { useUserStore, useReservationStore } from '@/store';
 import { storeToRefs } from 'pinia';
 import { useToast } from 'primevue/usetoast';
 import HeaderPanel from '@/components/inventory-list/HeaderPanel.vue';
 import ProfileName from '@/components/inventory-list/ProfileName.vue';
+import {
+  formatDate,
+  initFilters,
+  fetchData,
+  clearFilters,
+  searchItems,
+} from '@/utils/inventory';
 
 const userStore = useUserStore();
 const reservationStore = useReservationStore();
@@ -173,122 +179,21 @@ const requestees = ref([]);
 
 // set the default filter operators and constraints
 const filters = ref();
-
-function initFilters() {
-  filters.value = {
-    global: { value: null, matchMode: FilterMatchMode.CONTAINS },
-    status: {
-      operator: FilterOperator.OR,
-      constraints: [{ value: null, matchMode: FilterMatchMode.EQUALS }],
-    },
-    location: {
-      operator: FilterOperator.OR,
-      constraints: [{ value: null, matchMode: FilterMatchMode.EQUALS }],
-    },
-    startDate: {
-      operator: FilterOperator.AND,
-      constraints: [{ value: null, matchMode: FilterMatchMode.DATE_AFTER }],
-    },
-    endDate: {
-      operator: FilterOperator.AND,
-      constraints: [{ value: null, matchMode: FilterMatchMode.DATE_BEFORE }],
-    },
-    category: {
-      operator: FilterOperator.OR,
-      constraints: [{ value: null, matchMode: FilterMatchMode.IN }],
-    },
-    name: {
-      operator: FilterOperator.AND,
-      constraints: [{ value: null, matchMode: FilterMatchMode.STARTS_WITH }],
-    },
-    description: {
-      operator: FilterOperator.OR,
-      constraints: [{ value: null, matchMode: FilterMatchMode.CONTAINS }],
-    },
-    requestedBy: {
-      operator: FilterOperator.OR,
-      constraints: [{ value: null, matchMode: FilterMatchMode.EQUALS }],
-    },
-    requestedDate: {
-      operator: FilterOperator.AND,
-      constraints: [{ value: null, matchMode: FilterMatchMode.DATE_AFTER }],
-    },
-  };
-}
-
-// create a list of unique filter options
-function createUniqueFilterOptions(data, field) {
-  return [...new Set(data.map((item) => item[field]))];
-}
-
-// set the filter options for the dropdowns
-function setFilterOptions(newReservations) {
-  const lists = {
-    status: statuses,
-    location: locations,
-    category: categories,
-    requestedBy: requestees,
-  };
-
-  Object.keys(lists).forEach((key) => {
-    lists[key].value = createUniqueFilterOptions(newReservations, key);
-  });
-
-  // statuses.value = createUniqueFilterOptions(newReservations, 'status');
-  // locations.value = createUniqueFilterOptions(newReservations, 'location');
-  // categories.value = createUniqueFilterOptions(newReservations, 'category');
-  // requestees.value = createUniqueFilterOptions(newReservations, 'requestedBy');
-}
-
-// clear the filters
-const clearFilter = () => {
-  initFilters();
+const clearFilter = () => clearFilters(filters);
+const lists = {
+  status: statuses,
+  location: locations,
+  category: categories,
+  requestedBy: requestees,
 };
-
-// fetch the data from the server or test data file
-// and set the reactive variables
-async function fetchData() {
-  loading.value = true;
-  try {
-    await reservationStore.fetchReservations();
-    setFilterOptions(reservationStore.reservations);
-  } catch (err) {
-    console.error(err);
-  }
-  loading.value = false;
-}
-
-// format the date to a readable format
-function formatDate(date) {
-  return date.toLocaleString('en-US', {
-    month: 'short',
-    day: 'numeric',
-    year: 'numeric',
-    hour: 'numeric',
-    minute: 'numeric',
-  });
-}
 
 // create a list of item names for the autocomplete
 const itemNames = computed(() => reservationStore.reservations.map((item) => item.name));
 const filteredItems = ref([]);
 
-const searchItems = (event) => {
-  // in final app, make request to db with query and return filtered results
-  // for testing filter at client side
-  const { query } = event;
-  const filteredItemsList = [];
-
-  for (let i = 0; i < itemNames.value.length; i += 1) {
-    const item = itemNames.value[i];
-
-    if (item.toLowerCase().indexOf(query.toLowerCase()) === 0) {
-      filteredItemsList.push(item);
-    }
-  }
-
-  filteredItems.value = filteredItemsList;
-};
+function search(event) {
+  searchItems(event, itemNames, filteredItems);
+}
 
 function approveReservation(data) {
   reservationStore.approveReservation(data.id);
@@ -310,6 +215,11 @@ function denyReservation(data) {
   });
 }
 
+async function requestData() {
+  const fetchFn = reservationStore.fetchReservations;
+  await fetchData(fetchFn, reservationStore.reservations, loading, lists);
+}
+
 const router = useRouter();
 
 // redirect to home if not admin
@@ -317,11 +227,11 @@ onBeforeMount(() => {
   if (!isAdmin.value) {
     router.push('/');
   }
-  initFilters();
+  initFilters(filters);
 });
 // fetch data when the view is created
 onMounted(async () => {
-  await fetchData();
+  await requestData();
 });
 </script>
